Converted part of addToFavorites to Python
It was required to convert this function as part of the Python migration
process. However for the moment only the part that handles the files
update has been converted. The part that provides the function argument
consisting of the absolute directory path remains in BASH at the time.
It is planned to have it added to Python at a later stage