Refactored the executeCommand code
It was required to move the files update code at the end of the function
to help solve a bug that appeared in some Linux Systems (like OpenSUSE
15.1) that caused the persistent command history to get hidden after the execution of any
command. The bug is not entirely solved but the situations when it
appears have been drastically reduced. It is probably related to Python
library installed on the affected OSes as on other systems (like Mac OS and Fedora) no such issue has been encountered so far