Fixed a bug that appears both on Linux and Mac in different scenarios
The bug consisted of vanishing persistent command history entries when displaying the command history menu. The circumstances are different among OSes: on Linux the entries vanish after executing a command (either from commands history menu or a manually entered command). On Mac the entries dissappear after opening a new terminal window. The root causes could not be determined but this fix (not ideal but best that could be found) solves the issue