Corrected some loops from NavigationBackend where list elements are being removed
This was required in order to solve a bug related to incorrect updating of daily log. Although untracked, potential bugs might have existed in other sections of the code as well. For this reason a copy of each list needed to be created for iteration purposes while the actual removal would occur within original list