Refactored the general output, help and close application messages display
It was required to move all current dir sync commands from display.py. This was needed in order to ensure this module only deals with displaying data and this enforce the single responsibility principle

Also it was required to remove an unnecessary function from display.py that was used in the past for displaying quick history. This function is no longer required as the functionality is implemented by a method from goto_app.py